Add ball movement ,collisions, and losing conditions
ball moves randomly. 
If collide with top or bottom wall, it will bounce off in the opposite direction in y axis.
If collide with racket, it will bounce off in the opposite direction in x axis.
If collide with side walls, it will trigger a losing screen (Text depending on which side it collided with).